raft: group0_tables: add definition of intermediate language
In group0 tables, raft command contains a whole program to be executed.
Sending and parsing on each node entire CQL statement is inefficient,
thus we decided to compile it to an intermediate language which can be easily
serializable.

This patch adds a definition of such a language. For now, only the following
types of statements can be compiled:
* select value where key = CONST from system.group0_kv_store;
* update system.group0_kv_store set value = CONST where key = CONST;
* update system.group0_kv_store set value = CONST where key = CONST if value = CONST;
where CONST is string literal.