raft: group0_tables: add compilation of cql to intermediate language
We decided to add execution path of queries on group0 tables after it's
compiled to cql_statement. It's dispatched in query_processor::process_authorized_statement
instead of normal path if GROUP0_TABLES flag is enabled and it operates
on system.group0_kv_store.

If query is executed on shard other than 0 it's bounced to that shard.