raft topology: non-node-specific `topology_mutation_builder`
The existing `topology_mutation_builder` took a `raft::server_id` in its
constructor and immediately created a clustering row in the
`system.topology` mutation that it was building for the given node.
This does not allow building mutations which only affect the static
columns.

Split the class into two:
- `topology_mutation_builder` doesn't take `raft::server_id` in its
  constructor and contains only the methods that are used to set static
  columns. It also has a `with_node` method taking a `raft::server_id`
  which returns a `topology_node_mutation_builder&`.
- `topology_node_mutation_builder` creates the clustering row and allows
  seting its columns.

We'll use `topology_mutation_builder` when we only want to transition
the cluster-global topology state, without affecting any specific nodes'
states.