raft topology: introduce `start_operation()`
This calls `raft_group0_client::start_operation` and checks if the term
is different from the term that the coordinator was initially created
with; if so, we must no longer continue coordinating the topology.

There was one direct call to `raft_group0_client::start_operation`
without a term check, replace it with the introduced function.