raft topology: introduce `prepare_and_broadcast_cdc_generation_data`
This broadcasts all mutations returned from `prepare_new_cdc_generation_data`
except the last one. Each mutation is sent in separate raft command and is
retried as long as `group0_concurrent_modification` occurs.
It takes `group0_guard` and if the number of mutations is greater than one,
the guard is dropped and a new one is created and returned, otherwise the
old one will be returned. Commands are sent sequentially and unguarded (guard
used for sending the last mutation will guarantee that the term hasn't been changed).
Returns the generation's UUID, guard and last mutation which will be sent
with additional topology data by the caller.

With the default commitlog segment size, `mutation_size_threshold` will be 4 MB.
In large clusters e.g. 100 nodes, 64 shards per node, 256 vnodes cdc generation
data can reach the size of 30 MB, thus there will be no more than 8 commands.

In a multi-DC cluster with 100ms latencies between DCs, which means ~200ms for
each command, because we first perform a barrier in start_operation and then commit
the command. Thus this operation will take about 1.6s which is incomparably shorter
than bootstrapping operation (bootstrapping is quick if there is no data in the cluster,
but usually if one has 100 nodes they have tons of data, so indeed streaming/repair
will take much longer (hours/days)).

The alternative solution to sequential sending could be sending these commands
concurrently. But it would be complex, it's not clear how to do it with `group0_guard`
and this operation's time is not a problem.

Fixes FIXME in pr #13683.