raft: stop group0 server during group0 service shutdown
When the `topology_change` command is applied, the topology state is reloaded
and `cdc::generation_service::handle_cdc_generation` is called. This creates
a dependency of the group0 on `cdc::generation_service`.
Currently, the group0 server is stopped during `raft_group_registry` shutdown.
However, it is called after `cdc::generation_service` shutdown, which can
result in a segfault.

To prevent this issue, this commit stops the group0 server and removes it
from the `raft_group_registry` during the `group0_service` shutdown.

Fixes #14397.