main: add maintenance mode
In maintenance mode:
* Group0 doesn't start and the node doesn't join the token ring to behave as a dead
node to others,
* Group0 operations are disabled and result in an error,
* Only the maintenance port listens for CQL requests. This port allows all incoming
queries, so it should listen on the localhost address,
* The storage service initialises token_metadata with the local node as the only
node on the token ring.

Maintenance mode is enabled by passing the --maintenance-mode flag.
Flag --enable-maintenance-port also needs to be enabled for maintenance mode.

Maintenance mode starts before the group0 is initialised. This is done by moving
the initialisation of the cql server controllers and auth services earlier.