cql_controller: add option to start maintenance socket
Add an option to listen on the maintenance socket. It is set up on a separate
listen address and metrics are disabled.
This enables having an independent authentication mechanism for this port.

To start the maintenance socket, a new cql_controller has to be created
with
`db::maintenance_socket_enabled::yes` argument.

The indentation is fixed in the next commit.